Added Creation Functionality
HTML could use a little modification to make it more user friendly
Currently, if user goes to creation view and inputs nothing, then the page will not change. Most likely will change that in the future.

+++ WellnessClimbingWall/Models/Route.cs
@@ -10,7 +10,7 @@ namespace WellnessClimbingWall.Models
         public int RouteId { get; set; }
         public string Grade { get; set; }
         public string Color { get; set; }
-        public int Rope { get; set; }
+        public ushort Rope { get; set; }
         public DateTime DateCreated { get; set; }
         public string Location { get; set; }
         public string Setter { get; set; }
